Profiling and some small optimization

+++ flatland_plugins/include/flatland_plugins/laser.h
@@ -73,12 +73,12 @@ class Laser : public ModelPlugin {
   std::string topic_;     ///< topic name to publish the laser scan
   Body *body_;            ///<  body the laser frame attaches to
   Pose origin_;           ///< laser frame w.r.t the body
-  double range_;          ///< laser max range
-  double noise_std_dev_;  ///< noise std deviation
-  double max_angle_;      /// < laser max angle
-  double min_angle_;      ///< laser min angle
-  double increment_;      ///< laser angle increment
-  double update_rate_;    ///< the rate laser scan will be published
+  float range_;           ///< laser max range
+  float noise_std_dev_;   ///< noise std deviation
+  float max_angle_;       /// < laser max angle
+  float min_angle_;       ///< laser min angle
+  float increment_;       ///< laser angle increment
+  float update_rate_;     ///< the rate laser scan will be published
   std::string frame_id_;  ///< laser frame id name
   bool broadcast_tf_;     ///< whether to broadcast laser origin w.r.t body
   bool flipped_;          ///< whether the lidar is flipped
@@ -91,8 +91,8 @@ class Laser : public ModelPlugin {
    */
   uint16_t reflectance_layers_bits_;
 
-  std::default_random_engine rng_;              ///< random generator
-  std::normal_distribution<double> noise_gen_;  ///< gaussian noise generator
+  std::default_random_engine rng_;             ///< random generator
+  std::normal_distribution<float> noise_gen_;  ///< gaussian noise generator
 
   Eigen::Matrix3f m_body_to_laser_;       ///< tf from body to laser
   Eigen::Matrix3f m_world_to_body_;       ///< tf  from world to body
